Created empty interface Drawable and AbstractTile. Moved Tile functions into AbstractTile
