+++ sailinggame/src/main/java/com/group11/view/AppWindow.java
@@ -34,7 +34,7 @@ public class AppWindow {
 
     /**
      * Testing purposes. Will be removed later.
-     * @return {List<List<Integer>>} Matrix of integers corresponding to terrain id:s
+     * @return Matrix of integers corresponding to terrain id:s
      */
     private List<List<Integer>> initializeTerrainTest() {
         List<List<Integer>> testTerrainList = new ArrayList<>();
@@ -62,9 +62,9 @@ public class AppWindow {
      * Adds each JLabel component, obtained from the Tile object, on the map to the main window frame
      */
     private void addTilesToFrame() {
-        ArrayList<ArrayList<Tile>> tileMatrix = this.gameWorld.getTileMatrix();
-        for (ArrayList<Tile> tileRow : tileMatrix) {
-            for (Tile tile : tileRow) {
+        ArrayList<ArrayList<AbstractTile>> tileMatrix = this.gameWorld.getTileMatrix();
+        for (ArrayList<AbstractTile> tileRow : tileMatrix) {
+            for (AbstractTile tile : tileRow) {
                 mainFrame.add(tile.getComponent());
             }
         }
