+++ sailinggame/src/main/java/com/group11/view/Drawable.java
@@ -1,0 +1,5 @@
+package com.group11.view;
+
+public interface Drawable {
+
+}
